+++ packages/stream_chat/lib/src/client/channel.dart
@@ -3303,9 +3303,9 @@ class ChannelClientState {
   List<Read> get read => _channelState.read ?? <Read>[];
 
   /// Channel read list as a stream.
-  Stream<List<Read>> get readStream {
-    return channelStateStream.map((cs) => cs.read ?? <Read>[]);
-  }
+  Stream<List<Read>> get readStream => channelStateStream
+      .map((cs) => cs.read ?? <Read>[])
+      .distinct(const ListEquality().equals);
 
   /// Channel read for the logged in user.
   Read? get currentUserRead {
@@ -3315,14 +3315,16 @@ class ChannelClientState {
 
   /// Channel read for the logged in user as a stream.
   Stream<Read?> get currentUserReadStream {
-    final currentUser = _client.state.currentUserStream;
-    return currentUser.switchMap((it) => userReadStreamOf(userId: it?.id));
+    final currentUser =
+        _client.state.currentUserStream.mapNotNull((it) => it?.id).distinct();
+    return currentUser
+        .switchMap((id) => userReadStreamOf(userId: id))
+        .distinct();
   }
 
   /// Unread count getter as a stream.
-  Stream<int> get unreadCountStream {
-    return currentUserReadStream.map((read) => read?.unreadMessages ?? 0);
-  }
+  Stream<int> get unreadCountStream =>
+      currentUserReadStream.map((read) => read?.unreadMessages ?? 0).distinct();
 
   /// Unread count getter.
   int get unreadCount => currentUserRead?.unreadMessages ?? 0;