+++ packages/stream_chat_flutter/test/scrollable_positioned_list/seperated_horizontal_scrollable_positioned_list_test.dart
@@ -78,8 +78,8 @@ void main() {
         itemScrollController: itemScrollController,
         itemPositionsListener: itemPositionsListener);
 
-    unawaited(
-        itemScrollController.scrollTo(index: 2, duration: scrollDuration));
+    unawaited(itemScrollController.scrollTo(
+        index: 2, duration: scrollDuration, curve: Curves.linear));
     await tester.pump();
     await tester.pump(scrollDuration);
 
@@ -108,8 +108,8 @@ void main() {
         itemScrollController: itemScrollController,
         itemPositionsListener: itemPositionsListener);
 
-    unawaited(
-        itemScrollController.scrollTo(index: 100, duration: scrollDuration));
+    unawaited(itemScrollController.scrollTo(
+        index: 100, duration: scrollDuration, curve: Curves.linear));
     await tester.pumpAndSettle();
 
     expect(find.text('Item 99'), findsNothing);
@@ -168,8 +168,8 @@ void main() {
     expect(tester.getBottomRight(find.text('Item 1')),
         const Offset(10 + itemWidth * 2 + separatorWidth, screenHeight - 10));
 
-    unawaited(
-        itemScrollController.scrollTo(index: 494, duration: scrollDuration));
+    unawaited(itemScrollController.scrollTo(
+        index: 494, duration: scrollDuration, curve: Curves.linear));
     await tester.pumpAndSettle();
 
     await tester.drag(